AMBARI-9016 BUG-105818 Fixed java package location and changed inner class visibility. - ContentTypeOverrideFilter was moved to java.org.apache.ambari.server.api package. - Inner Response/Request wrapper classes visibility has changed to package-private   to support unit testing.
