+++ ambari-server/src/test/java/org/apache/ambari/server/api/ContentTypeOverrideFilterTest.java
@@ -1,0 +1,4 @@
+package org.apache.ambari.server.api;
+
+public class ContentTypeOverrideFilterTest {
+}
